pdf-converter-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdf2image==1.17.0
Pillow==10.1.0
python-magic==0.4.27 